requirements.txt
requests
xterm
pillow
customtkinter